Feature/add map

+++ components/Event/index.js
@@ -1,7 +1,9 @@
+import Link from "next/link";
 import { useRouter } from "next/router";
 import { useState } from "react";
 import styled from "styled-components";
 import AddEventForm from "../AddEventForm";
+import SVGIcon from "../Icons";
 
 const EventContainer = styled.article`
   border: 2px solid black;
@@ -10,27 +12,53 @@ const EventContainer = styled.article`
   justify-content: center;
   align-items: center;
   flex-direction: column;
-  background-color: #ccd9ff;
+  background-color: hsla(0, 0%, 4%, 0.64);
   border-radius: 2rem;
+  position: relative;
+  color: lightgray;
+  max-width: 300px;
 `;
 const Title = styled.p`
-  width: 80%;
+  width: 90%;
   text-align: center;
-  margin-bottom 0.4rem;
   font-weight: bold;
-  font-size: 1.2rem;
+  font-size: 1.4rem;
   overflow-wrap: break-word;
-
 `;
 const Date = styled.p`
   align-self: flex-start;
   padding-left: 1rem;
 `;
-const Description = styled.p`
+const StyledDescription = styled.p`
   overflow-wrap: break-word;
-  background-color: lightgray;
   height: fit-content;
   width: 15rem;
+  background-color: hsla(0, 0%, 100%, 0.22);
+  border: 3px solid black;
+  border-radius: 1rem;
+  padding: 1rem 1rem;
+`;
+const StyledEditButton = styled.button`
+  position: absolute;
+  top: -1.2rem;
+  right: 2.5rem;
+  background-color: hsla(34, 93%, 52%, 0.89);
+  border-radius: 999px;
+`;
+const StyledDeleteButton = styled.button`
+  position: absolute;
+  top: -1.2rem;
+  right: -0.4rem;
+  background-color: hsla(0, 93%, 40%, 0.89);
+  border-radius: 999px;
+`;
+const StyledLink = styled(Link)`
+  background-color: lightgray;
+  padding: 0.5rem 0.5rem;
+  height: 2rem;
+  border: 2px solid black;
+  border-radius: 1rem;
+  margin: 0.5 auto;
 `;
 
 export default function EventCard({ appointment }) {
@@ -38,6 +66,10 @@ export default function EventCard({ appointment }) {
   const [isEditing, setIsEditing] = useState(false);
 
   async function handleDelete() {
+    const sure = confirm("Do you really want to delete the document?");
+    if (!sure) {
+      return;
+    }
     try {
       const response = await fetch(`/api/appointments/${appointment._id}`, {
         method: "DELETE",
@@ -78,26 +110,24 @@ export default function EventCard({ appointment }) {
   }
   return (
     <EventContainer>
-      <button
+      <StyledEditButton
         type="button"
         onClick={() => {
           setIsEditing(!isEditing);
         }}
       >
-        Edit
-      </button>
+        <SVGIcon variant="edit" width="30px" />
+      </StyledEditButton>
+      <StyledDeleteButton type="button" onClick={handleDelete}>
+        <SVGIcon variant="delete" width="30px" />
+      </StyledDeleteButton>
       {isEditing ? (
-        <>
-          <button type="button" onClick={handleDelete}>
-            Delete
-          </button>
-          <AddEventForm onSubmit={handleSubmit} appointment={appointment} />
-        </>
+        <AddEventForm onSubmit={handleSubmit} appointment={appointment} />
       ) : appointment ? (
         <>
           <Title>{appointment.title}</Title>
           <Date>Date: {appointment.date}</Date>
-          <Description>{appointment.description}</Description>
+          <StyledDescription>{appointment.description}</StyledDescription>
           <p>
             Cost:{" "}
             {new Intl.NumberFormat("de-DE", {
@@ -106,6 +136,9 @@ export default function EventCard({ appointment }) {
             }).format(appointment.cost)}
           </p>
           <p>Priority: {appointment.priority}</p>
+          <StyledLink href={`/appointmentDetails/${appointment._id}`}>
+            details
+          </StyledLink>
         </>
       ) : null}
     </EventContainer>
